avformat/omadec: Use 64bit for ret to avoid overflow
(cherry picked from commit 0f55bc29d41585d110b126cb4ed4b395fd46d7ac)

+++ libavformat/omadec.c
@@ -464,7 +464,7 @@ static int oma_read_seek(struct AVFormatContext *s,
                          int stream_index, int64_t timestamp, int flags)
 {
     OMAContext *oc = s->priv_data;
-    int err = ff_pcm_read_seek(s, stream_index, timestamp, flags);
+    int64_t err = ff_pcm_read_seek(s, stream_index, timestamp, flags);
 
     if (!oc->encrypted)
         return err;
